Initial commit
I want to build a library for ITU-R BS.1770 loudness analysis, so I can
use it to do loudness normalization in Musium. There are already various
programs that can analyze loudness according to BS.1770 and write the
result in the form of ReplayGain. I want to do this manually mainly out
of personal interest.