Add De Man filter coefficient formulas
This adds the filter coefficients by De Man, which are a generalization
of the coefficients in the spec to arbitrary sample rates (the spec only
specifies coefficients for 48 kHz). Also add a test to confirm that the
coefficients match the spec.

The actual formulas are not taken from the work of De Man: the work is
inaccessible because it is behind a paywall. Instead, the formulas are
taken from Christian Steinmetz Python implementation of De Man's work,
available at https://github.com/csteinmetz1/pyloudnorm.