Add test case 7 and 8 from Tech 3341
These are the files with the reference material that can be downloaded
with the script I added previously. The test passes now, but I did
originally make a mistake in determining the maximum amplitude. This
mistake was also present in the main executable. After fixing it, my
meaturements agree to one decimal with bs1770gain, though there is a
slight difference. It might be that bs1770gain performs upstampling to
get a more accurate power measurement (I should still add that), or it
might be because I use compensated summing ... not sure.