Add some type safety
The Power type does not say anything about timescales, which is good.
But for the gated measurement, I want the 100ms non-overlapping windows.
Wrapping the windows ensures that by default, the pieces fit together
only in a single way. It also clarifies where reduce_stereo fits in.

The types are transparent to the user, so users can still construct
these outside of the library. It is not a correctness guarantee, but
rather a strong nudge that ensures the default thing to do is the
correct thing, without making advanced usage impossible for users who do
know what they are doing.