Move main program into examples directory
I want to be able to publish this on crates.io, and the library itself
does not have dependencies on Claxon or Hound. The main program does,
but it should not be the main part of the library. By moving it into the
examples directory, the dependencies of the main program can be
dev-dependencies.