Add script to run flacgain on an entire library
I want to add these tags to my library now. Parallelism is a must-have,
although the disk is the bottleneck at this point. I could do this in
Rust, but it is easier to just throw a few existing tools together;
maybe it is not as clear and robust, but it will do for now. I can
inspect the cases where it did not write any tags manually later.