Fix bug in enumerating leaf directories with find
This is what I get for not wanting to write a Python or Rust script to
do this, and then trusting a hack from the interwebs because find doen't
have a way to list leaves.

There was a bug where directories that are a prefix of a sibling
directory did not get matched. For example

    /music/Toto/Toto IV
    /music/Toto/Toto                  <-- Not matched
    /Music/Various Artists/Alchemy 2
    /Music/Various Artists/Alchemy    <-- Not matched

Fix this by not caring about whether a directory is a leaf or not. To
stop find from recursively finding files, add a -maxdepth of 1.